+++ templates/default/src/include/src/components/FeedItems.vue
@@ -1,15 +1,25 @@
 <script setup>
-import { ref, watchEffect, shallowRef } from 'vue'
-import { useFiltersStore } from '../stores/filters'
+import { ref, watchEffect, shallowRef, onMounted } from 'vue'
 import { RouterLink } from 'vue-router'
-import IconMusic from './icons/IconMusic.vue'
-import IconMovie from './icons/IconMovie.vue'
 import { useEmbedStore } from '../stores/embed'
 import { useAudioStore } from '../stores/audio'
+import { useFiltersStore } from '../stores/filters'
+import { useFeedItemsStore } from '../stores/feedItems'
+import { useMinimizeStore } from '../stores/minimize'
+import IconMusic from './icons/IconMusic.vue'
+import IconMovie from './icons/IconMovie.vue'
+import IconExpand from './icons/IconExpand.vue'
+import IconMinimize from './icons/IconMinimize.vue'
+import IconMaximize from './icons/IconMaximize.vue'
+import IconTop from './icons/IconTop.vue'
 
 const fStore = useFiltersStore()
 const embedStore = useEmbedStore()
 const audioStore = useAudioStore()
+const minimizeStore = useMinimizeStore()
+const fItemsStore = useFeedItemsStore()
+
+const { getFeedItems } = fItemsStore
 
 const props = defineProps({
   feed: {
@@ -24,6 +34,10 @@ const props = defineProps({
     type: Boolean,
     required: true,
   },
+  firehose: {
+    type: Boolean,
+    required: true,
+  },
   expand: {
     type: Boolean,
     required: true,
@@ -34,10 +48,10 @@ const props = defineProps({
   },
 })
 
-const feedItems = shallowRef([])
 const filteredFeedItems = shallowRef([])
 const initialized = ref(false)
 const emit = defineEmits(['expand-article', 'unexpand-article'])
+const itemDetails = ref(null)
 
 fStore.$subscribe((state) => {
   filterFeedItems(state.payload)
@@ -66,36 +80,6 @@ const truncate = (v) => {
 }
 // ===============
 
-
-const processFeedItems = (feedItems) => {
-  feedItems.sort((a, b) => {
-    return b.date - a.date
-  })
-
-  var result = []
-  for (let feedItem of feedItems) {
-    let date = new Date(feedItem.date * 1000)
-    let url
-    try {
-      url = new URL(feedItem.url)
-    } catch {
-      console.log('Could not fetch URL for article: ', feedItem)
-      continue
-    }
-    result.push({
-      title: feedItem.title,
-      url: feedItem.url,
-      date: date,
-      domain: url.hostname,
-      guid: feedItem.guid,
-      content: feedItem.content,
-      author: feedItem.author,
-      enclosureUrl: feedItem.enclosureUrl,
-    })
-  }
-  return result
-}
-
 const aggregateItems = (items) => {
   let result = {}
   let now = new Date()
@@ -112,51 +96,45 @@ const aggregateItems = (items) => {
   return result
 }
 
-const filterFeedItems = (state) => {
+const filterFeedItems = async (state) => {
+  const items = await retrieveItemData()
   if (state.searchTerm) {
-    filteredFeedItems.value = aggregateItems(_filterByTerm(state.searchTerm))
-  } else if (state.filterByDays) {
-    filteredFeedItems.value = aggregateItems(_updateItemsWithDate(state.daysBackCount))
-  } else {
-    filteredFeedItems.value = aggregateItems(_updateItemsWithCount(state.itemCount))
+    filteredFeedItems.value = aggregateItems(_filterByTerm(items, state.searchTerm))
+    return
   }
+  if (state.firehose) {
+    filteredFeedItems.value = aggregateItems(items)
+    return
+  }
+  if (state.filterByDays) {
+    filteredFeedItems.value = aggregateItems(_updateItemsWithDate(items, state.daysBackCount))
+    return
+  }
+  filteredFeedItems.value = aggregateItems(_updateItemsWithCount(items, state.itemCount))
 }
-const _filterByTerm = (term) => {
+const _filterByTerm = (items, term) => {
   let title = (props.feed.displayTitle || props.feed.title).toLowerCase().split(' ')
   let checker = (arr, target) => target.every((v) => arr.some((vv) => vv.includes(v)))
-  return feedItems.value.filter((f) => {
+  return items.filter((f) => {
     let fTitle = f.title.toLowerCase().split(' ')
     fTitle.push(f.author.toLowerCase())
     return checker(fTitle.concat(title), term.split(' '))
   })
 }
-const _updateItemsWithDate = (daysBack) => {
+const _updateItemsWithDate = (items, daysBack) => {
   let d = new Date()
   d.setDate(d.getDate() - daysBack)
-  return feedItems.value.filter((f) => {
+  return items.filter((f) => {
     return f.date > d
   })
 }
 
-const _updateItemsWithCount = (numItems) => {
-  return feedItems.value.slice(0, numItems)
+const _updateItemsWithCount = (items, numItems) => {
+  return items.slice(0, numItems)
 }
 
 const feedHasItems = () => {
   return Object.keys(filteredFeedItems.value).length !== 0
-}
-
-const resolveFeedPath = (feedId) => {
-  let basePath = `feeds/${feedId}`
-  if (props.archived) {
-    basePath = basePath + '_archive'
-  }
-  let pathPrefix = window.sitePath || '/'
-  if (!pathPrefix.endsWith('/')) {
-    pathPrefix = pathPrefix + '/'
-  }
-  let feedUrl = `${pathPrefix}${basePath}.json?bt=${window.buildTime}`
-  return feedUrl
 }
 
 // Feed/Article expansion
@@ -170,10 +148,11 @@ const handleExpandedArticle = (articleId) => {
 const handleUnexpandedArticle = (articleId) => {
   emit('unexpand-article', articleId)
 }
-const dispatchExpandItems = () => {
+const dispatchExpandItems = async () => {
+  const items = await retrieveItemData()
   return {
     feedId: props.feed.id,
-    articleIds: feedItems.value.map((i) => i.guid),
+    articleIds: items.map((i) => i.guid),
   }
 }
 // ======================
@@ -194,120 +173,161 @@ const showAudioPlayer = (feedItem) => {
   audioStore.setAudioData(feedItem)
   audioStore.showAudioPlayer()
 }
-// ===================
+const retrieveItemData = async () => {
+  return await getFeedItems(props.feed.id, props.archived)
+}
 
+// ===================
 watchEffect(async () => {
   if (!initialized.value) {
-    const url = resolveFeedPath(props.feed.id)
-    let data
-    try {
-      data = await (await fetch(url)).json()
-    } catch {
-      console.log('Could not fetch feed data for feed ', url)
-      return
-    }
-    feedItems.value = processFeedItems(data.items)
+    await retrieveItemData()
   }
   initialized.value = true
   if (props.filtered) {
     filterFeedItems(fStore)
   } else {
-    filteredFeedItems.value = aggregateItems(feedItems.value)
+    filteredFeedItems.value = aggregateItems(await retrieveItemData())
   }
+})
+
+onMounted(() => {
+  setInterval(() => {
+    if (!itemDetails.value || itemDetails.value.length === 0) {
+      return
+    }
+    const body = document.body
+    const docEl = document.documentElement
+
+    const scrollTop = window.pageYOffset || docEl.scrollTop || body.scrollTop
+    const clientTop = docEl.clientTop || body.clientTop || 0
+    const center = scrollTop + window.innerHeight / 2
+    let visibleDetails = []
+    for (let detail of itemDetails.value) {
+      const box = detail.getBoundingClientRect()
+      const boxTop = box.top + scrollTop - clientTop
+      const boxCenter = box.top + scrollTop - clientTop + box.height / 2
+      // filter out invisible details
+      if (boxTop + box.height < scrollTop || scrollTop + window.innerHeight < boxTop) {
+        detail.classList.remove('detail-highlight')
+      } else {
+        detail.setAttribute('offsetCenter', Math.abs(center - boxCenter).toString())
+        visibleDetails.push(detail)
+      }
+    }
+    if (visibleDetails.length === 0) {
+      return
+    }
+    visibleDetails.sort((a, b) => {
+      return parseFloat(a.getAttribute('offsetCenter')) - parseFloat(b.getAttribute('offsetCenter'))
+    })
+
+    const first = visibleDetails.shift()
+    first.classList.add('detail-highlight')
+    for (let d of visibleDetails) {
+      d.classList.remove('detail-highlight')
+    }
+  }, 400)
 })
 </script>
 
 <template>
   <div class="feed-wrapper" v-if="feedHasItems()">
     <div class="feed-title">
-      <router-link :to="{ name: 'feedView', params: { feedId: feed.id } }" v-if="feed.title"
+      <router-link :to="{ name: 'feedView', params: { feedId: feed.id } }" v-if="!props.firehose"
         >{{ feed.displayTitle || feed.title }}
         <span v-if="feed.isQuery" class="feed-query-indicator"></span>
         <span class="item-count">({{ feed.itemCount }})</span></router-link
       >
-      <button
-        @click="$emit('expand-feed', dispatchExpandItems())"
-        class="expand-button"
-        title="Expand"
-        v-if="!props.expand && !props.archived"
-      >
-        &#8675;
-      </button>
-      <button
-        @click="$emit('unexpand-feed')"
-        class="expand-button"
-        title="Unexpand"
-        v-if="props.expand && !props.archived"
-      >
-        &#8673;
-      </button>
+      <a v-else href="#">{{ feed.displayTitle }}</a>
+      <span class="feed-buttons">
+        <button
+          @click="minimizeStore.addMinimizedFeed(feed.id)"
+          class="minimize-button"
+          title="Minimize"
+          v-if="!minimizeStore.showFeedMinimized(feed.id) && !props.archived && !props.firehose"
+        >
+          <IconMinimize />
+        </button>
+        <button
+          @click="minimizeStore.removeMinimizedFeed(feed.id)"
+          class="minimize-button"
+          title="Maximize"
+          v-if="minimizeStore.showFeedMinimized(feed.id) && !props.archived && !props.firehose"
+        >
+          <IconMaximize />
+        </button>
+        <button
+          @click="$emit('expand-feed', dispatchExpandItems())"
+          class="expand-button feed-expand-button"
+          title="Expand"
+          v-if="!props.expand && !props.archived && !props.firehose"
+        >
+          <IconTop />
+        </button>
+        <button
+          @click="$emit('unexpand-feed')"
+          class="expand-button feed-unexpand-button"
+          title="Unexpand"
+          v-if="props.expand && !props.archived && !props.firehose"
+        >
+          <IconTop />
+        </button>
+      </span>
     </div>
-    <div class="feed-item-group" v-for="(items, dateStr) in filteredFeedItems" :key="dateStr">
-      <span class="feed-group-date" v-if="dateStr">{{ dateStr }}</span>
-      <TransitionGroup name="items" tag="ul">
-        <li v-for="(feedItem, index) in items" :key="index" class="feed-item">
-          <ArticleItem
-            :feedItem="feedItem"
-            :expand="showExpandedArticle(feedItem.guid)"
-            @expand-article="handleExpandedArticle(feedItem.guid)"
-            @unexpand-article="handleUnexpandedArticle(feedItem.guid)"
-          />
-
-          <span class="feed-item-link">
-            <a
-              v-if="embedStore.isEmbeddable(feedItem)"
-              @click="showEmbedModal(feedItem)"
-              target="_blank"
+    <div v-if="!minimizeStore.showFeedMinimized(feed.id)">
+      <div class="feed-item-group" v-for="(items, dateStr) in filteredFeedItems" :key="dateStr">
+        <span class="feed-group-date" v-if="dateStr">{{ dateStr }}</span>
+        <TransitionGroup name="items" tag="ul">
+          <li v-for="(feedItem, index) in items" :key="index" class="feed-item">
+            <span class="feed-item-link">
+              <a
+                v-if="embedStore.isEmbeddable(feedItem)"
+                @click="showEmbedModal(feedItem)"
+                target="_blank"
+              >
+                {{ truncate(feedItem.title) }}<span class="feed-item-type"><IconMovie /></span
+              ></a>
+              <a
+                v-else-if="audioStore.isAudioLink(feedItem)"
+                @click="showAudioPlayer(feedItem)"
+                target="_blank"
+              >
+                {{ truncate(feedItem.title) }}<span class="feed-item-type"><IconMusic /></span>
+              </a>
+              <a v-else :href="feedItem.url" target="_blank">{{ truncate(feedItem.title) }}</a>
+              <button
+                @click="handleExpandedArticle(feedItem.guid)"
+                class="expand-button article-expand"
+                title="Expand"
+                v-if="!showExpandedArticle(feedItem.guid)"
+              >
+                <IconExpand />
+              </button>
+              <button
+                @click="handleUnexpandedArticle(feedItem.guid)"
+                class="expand-button article-expand article-unexpand"
+                title="Unexpand"
+                v-if="showExpandedArticle(feedItem.guid)"
+              >
+                <IconExpand />
+              </button>
+            </span>
+            <span class="feed-item-author" v-if="feedItem.author"> by {{ feedItem.author }}</span>
+            <span class="feed-item-domain">({{ feedItem.domain }})</span>
+            <div
+              :class="{ 'feed-item-details': true, expanded: showExpandedArticle(feedItem.guid) }"
+              v-if="showExpandedArticle(feedItem.guid) && feedItem.content"
+              ref="itemDetails"
             >
-              {{ truncate(feedItem.title) }}<span class="feed-item-type"><IconMovie /></span
-            ></a>
-            <a
-              v-else-if="audioStore.isAudioLink(feedItem)"
-              @click="showAudioPlayer(feedItem)"
-              target="_blank"
-            >
-              {{ truncate(feedItem.title) }}<span class="feed-item-type"><IconMusic /></span>
-            </a>
-            <a v-else :href="feedItem.url" target="_blank">{{ truncate(feedItem.title) }}</a>
-            <button
-              @click="handleExpandedArticle(feedItem.guid)"
-              class="expand-button article-expand"
-              title="Expand"
-              v-if="!showExpandedArticle(feedItem.guid)"
-            >
-              &#8675;
-            </button>
-            <button
-              @click="handleUnexpandedArticle(feedItem.guid)"
-              class="expand-button article-expand"
-              title="Unexpand"
-              v-if="showExpandedArticle(feedItem.guid)"
-            >
-              &#8673;
-            </button>
-          </span>
-          <span class="feed-item-author" v-if="feedItem.author"> by {{ feedItem.author }}</span>
-          <span class="feed-item-domain">({{ feedItem.domain }})</span>
-          <div
-            :class="{ 'feed-item-details': true, expanded: showExpandedArticle(feedItem.guid) }"
-            v-if="showExpandedArticle(feedItem.guid)"
-          >
-            <span class="feed-item-date"
-              ><span class="feed-item-details-desc">Date: </span
-              >{{ feedItem.date.toUTCString() }}</span
-            >
-            <br />
-
-            <span class="feed-item-url"
-              ><span class="feed-item-details-desc">URL: </span>{{ feedItem.url }}</span
-            ><br />
-            <span class="feed-item-contents" v-if="feedItem.content"
-              ><span class="feed-item-details-desc">Content: </span>
-              <span v-html="feedItem.content"></span
-            ></span>
-          </div>
-        </li>
-      </TransitionGroup>
+              <span class="feed-item-contents"
+                ><span class="feed-item-details-desc">-------</span><br />
+                <span v-html="feedItem.content"></span><br />
+                <span class="feed-item-details-desc">--------</span>
+              </span>
+            </div>
+          </li>
+        </TransitionGroup>
+      </div>
     </div>
   </div>
 </template>
@@ -317,7 +337,6 @@ watchEffect(async () => {
   display: inline-block;
   height: 18px;
   width: 18px;
-  background-color: var(--color-custom);
   position: relative;
   top: 1px;
   margin: 0 4px;
@@ -330,6 +349,7 @@ watchEffect(async () => {
   transform: translateY(-10%);
   font-weight: bold;
   font-size: 0.8rem;
+  color: var(--color-custom);
 }
 .item-count {
   opacity: 0.6;
@@ -372,18 +392,21 @@ watchEffect(async () => {
 }
 .feed-item-details {
   opacity: 0.8;
-  outline: 1px solid rgb(from var(--color-highlight) r g b / 60%);
-  padding: 20px;
-  border-radius: 10px;
+  padding: 40px;
   overflow: hidden;
   display: none;
 }
+
 .feed-item-details-desc {
   color: var(--color-highlight);
 }
 .feed-item-details.expanded {
   display: block;
 }
+.feed-item-details.detail-highlight {
+  opacity: 1;
+}
+
 .feed-item-domain {
   opacity: 0.4;
   font-size: 0.72rem;
@@ -417,10 +440,10 @@ watchEffect(async () => {
   opacity: 0.7;
 }
 
-.expand-button {
+.expand-button,
+.minimize-button {
   display: inline-block;
   position: relative;
-  top: 1px;
   cursor: pointer;
   border: none;
   background: transparent;
@@ -428,8 +451,36 @@ watchEffect(async () => {
   color: var(--color-text);
   font-size: 1.2rem;
 }
+.minimize-button svg {
+  width: 20px;
+  height: 20px;
+}
+
+.feed-expand-button,
+.feed-unexpand-button {
+  color: var(--color-custom);
+  opacity: 0.8;
+  top: -2px;
+}
+.feed-expand-button svg {
+  transform: rotate(180deg);
+}
 .expand-button:hover {
   opacity: 1;
+}
+.article-expand {
+  top: 1px;
+}
+.article-expand svg {
+  width: 18px;
+  height: 18px;
+  position: relative;
+  top: 4px;
+  color: var(--color-text);
+  opacity: 0.8;
+}
+.article-unexpand svg {
+  transform: rotate(90deg);
 }
 
 @media (min-width: 1150px) {